refactor: move role creation logic to CreateRoleService and simplify RoleController
